[refactor] modified file structure
