+++ frontend/app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name="com.project.spire.ui.create.image.WriteTextActivity"
+            android:name="com.project.spire.ui.create.WriteTextActivity"
             android:exported="true" />
         <activity
             android:name="com.project.spire.ui.auth.VerifyEmailActivity"
@@ -33,7 +33,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name="com.project.spire.ui.create.image.ImageEditActivity"
+            android:name="com.project.spire.ui.create.ImageEditActivity"
             android:exported="true"
             android:label="@string/title_activity_image_edit" />
         <activity
