+++ frontend/app/src/main/java/com/project/spire/network/inference/InferenceClient.kt
@@ -10,7 +10,7 @@ import retrofit2.create
 // Should be combined with RetrofitClient later
 class InferenceClient {
     companion object {
-        private const val BASE_URL = "http://localhost:8000/v2/"
+        private const val BASE_URL = "http://localhost:8080/v2/"
 
         private val okHttpClient = OkHttpClient.Builder()
             .build()
